+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 0.9.3
+* [#284](https://github.com/awakesecurity/proto3-suite/pull/284) Delete Repeated.hs
+  * [BREAKING CHANGE in experimental module: Rename `Prefix` to `FieldsEncoder`, along with related functions.]
+  * [BREAKING CHANGE in experimental module: Rename `Fields` to `FieldsEncoding`, along with related functions.]
+  * Add instances of `FromJSON`, `FromJSONPB`, `ToJSON`, `ToJSONPB`
+    for `MessageEncoding` and `MessageEncoder`.
+
+# 0.9.2
+* [#282](https://github.com/awakesecurity/proto3-suite/pull/282) Show MessageEncoder
+  * [BREAKING CHANGE in experimental module: Rename `toLazyByteString` to `messageEncoderToLazyByteString`.]
+  * [BREAKING CHANGE in experimental module: Rename `cacheMessage` to `messageCache`.]
+  * [BREAKING CHANGE in experimental module: Delete `Eq` instance for `MessageEncoding` because it is application-defined whether we should ignore field order during comparison.]
+  * Add `Show` instance for `MessageEncoder`.
+  * Add `messageEncoderToByteString` and `unsafeByteStringToMessageEncoder`.
+* [#283](https://github.com/awakesecurity/proto3-suite/pull/283) Delete Repeated.hs
+  * Delete unused experimental source file `src/Proto3/Suite/Form/Encode/Repeated.hs`;
+    proto3-wire provides the relevant functionality in `Proto3.Wire.Encode.Repeated`.
+
 # 0.9.1
 * [#275](https://github.com/awakesecurity/proto3-suite/pull/275) the `canonicalize-proto-file` executable no longer depends on the [`range-set-list](https://github.com/phadej/range-set-list#readme) package for normalizing reserved field ranges.
 * Relocated orphaned [`Pretty`](https://hackage.haskell.org/package/pretty-1.1.3.6/docs/Text-PrettyPrint-HughesPJClass.html#t:Pretty) instances for AST types to [`Proto3.Suite.DotProto.AST`].